Arreglado problema del float, era un problema de precisión. Mejor usaremos doubles ... Actualizado el simulador para que trabaje cómo una cola M/M/1.
git-svn-id: svn://localhost/tmp/simuib08/trunk@14 2480690e-8b93-4946-9c88-c6c64152b314

+++ src/systypes.h
@@ -10,12 +10,11 @@
 #include <stdbool.h>
 
 // Debug
-bool DEBUG; 
+bool DEBUG;
 
 // Tiempo
-typedef float TIME;
+typedef double TIME;
 #define NULLTIME -1.0
 #define INITIALTIME 0.0
-
 
 #endif // _SYSTYPES_H_
